+++ CHANGES.md
@@ -1,6 +1,6 @@
 0.6.3 (YYYY-MM-DD)
 ------------------
-- [ ] evaluate random model
+- [x] evaluate random model
 - [ ] evaluate label average model
 
 0.6.2 (2019-11-30)